Ver. 1.2.1: HTML5 version of Facebook Like button
The plugin now includes an option to use the HTML5 version of the
Facebook Like button. This allows users to add an individual text to the
timeline posting which is created by the Like button. A Facebook app ID
is required to use this option. For details see
socialshareprivacy-xl-readme.txt.